Add support for pretokenized and presgmented input text
